back/requirements.txt
flask-cors
flask==3.0.2
kafka-python==2.0.2